Commit compile_for_mac.sh
I wrote this script to make the work of compiling imgbrd-grabber on mac
easier to do.

It takes out a lot of the manual labour found in the compile doc:
	https://github.com/Bionus/imgbrd-grabber/wiki/Compilation#macos

Instead, it attempts to do the following, covering off as many prereqs
and bases as I could think:

1) If HomeBrew isn't installed, run the ruby script to install it
2) Install prerequisites
	cmake,
	gcc,
	qt5
    NOTE: The default qt5 does not have mysql libs compiled into it.
    If you need mysql, dialogues allow you to trigger brew into
    doing a full build of Qt...after a warning that this can take up to
    an hour or more to complete just on it's own.

3) It also installs a script dependancy: jq(1). It's a light-weight
commandline json parsing utility, and is used to parse Brew package
information cleanly.

4) Since imgbrd-grabber needs https, and build depends on openssl libs,
this script checks to see if any verison of openssl has been installed
with brew, and if not it also pulls in a bottled version with brew that
		//shouldn't//
clobber other openssl environments on the machine. I dont think it will
at least.

5) With all those prereqeuisites out of the way, it does a git pull to
update the gir repository.

If the script is not actually WITH the git respositry, the script clones
the repo from github into the location the script was executed from.

6) After that, the script finally starts doing the compile, mostly
according to the instructions at wiki/Compilation#macos

cmake has dependancies for OpenSSL though, so the script hooks it into
the bottle it downloaded or any other preexisting brew-installed openssl
libs it found before.

7) One difference the script does however is sort and move finished assets
arround into a mac .app folder so it can be run from the GUI.

Next to last is the script checks if there's a previous install of
imgbrd-grabber at ${HOME}/Applications/Grabber.app

If there IS an old version there, and the user says ok, then it starts
pulling user configuration data out of

	${HOME}/Applications/Grabber.app/Contents/MacOS
		/*.txt
		/*.ini
		/*.log
		/*.igl
		/thumbs/

and putting it all into the new build.

8) Finally the script lets the user put the new application into
${HOME}/Applications and launch it, with the preserved session from the
old app intact.