Handle brew versions of deps better; save last build instead of overwite
Handle the way the script detects versions of openssl and jq. Use a
different path in the json object to always select the last array element
(and thus the most recent version) of the package installed.

Use date-stamping to move previous installs of Grabber. If the build
went wrong, there is now a previous one left behind that can be failed
back to.