Fix issue with dangling pointer when calling utime (#2204)
Fix issue with dangling pointer when calling utime (fix #2203)

The temporary string object returned by the function call is
deallocated at the end of the expression. Since c_str() returns a
pointer to the internal string representation, the data is no longer
valid once the string object is destroyed. Due to this utime returns
ENOENT and Grabber prints an error in the log when modifying
file creation time.